refactor: move context in helpers folder
personally think context should not have its own folder, so moved it
into the helpers folder to make it more clear and so the app structure
has more sence

also ofc updating the path in the pages and components using context
